Change background-color, round the border of Board
Decided to go with a summer-ish color palette to the page, and rounded
the corners of the Board. This initially was a problem however, because
when the squares were drawn in, they would overflow THROUGH the rounded
corners. This was fixed by setting the property {overflow:hidden} to the
square container (.drawing-board)